+++ skbio/diversity/alpha/_base.py
@@ -7,6 +7,7 @@
 # ----------------------------------------------------------------------------
 
 from warnings import warn
+import functools
 
 import numpy as np
 from scipy.special import gammaln
@@ -17,6 +18,56 @@ from skbio.diversity._util import _validate_counts_vector
 from skbio.util._warning import _warn_deprecated
 
 
+def _validate_alpha(empty=None, cast_int=False):
+    """Validate counts vector for an alpha diversity metric.
+
+    Parameters
+    ----------
+    func : callable
+        Function that calculates an alpha diversity metric.
+    empty : any, optional
+        Return this value if set instead of calling the function when an input
+        community is empty (i.e., no taxon, or all taxa have zero counts).
+    cast_int : bool, optional
+        Cast values into integers, if not already. ``False`` by default.
+
+    Returns
+    -------
+    callable
+        Decorated function.
+
+    Notes
+    -----
+    This function serves as a decorator for individual functions that calculate
+    alpha diversity metrics. The first positional argument of a decorated
+    function must be a 1-D vector of counts/abundances of taxa in a community.
+    Additional arguments may follow.
+
+    """
+
+    def decorator(func):
+        @functools.wraps(func)
+        def wrapper(counts, *args, **kwargs):
+            counts = _validate_counts_vector(counts, cast_int)
+
+            # drop zero values, as these represent taxa that are absent from
+            # the community
+            if not (nonzero := counts != 0).all():
+                counts = counts[nonzero]
+
+            # return a value if community is empty (after dropping zeros)
+            if empty is not None and counts.size == 0:
+                return empty
+
+            # call function to calculate alpha diversity metric
+            return func(counts, *args, **kwargs)
+
+        return wrapper
+
+    return decorator
+
+
+@_validate_alpha(empty=np.nan)
 def berger_parker_d(counts):
     r"""Calculate Berger-Parker dominance index.
 
@@ -38,7 +89,7 @@ def berger_parker_d(counts):
 
     Returns
     -------
-    double
+    float
         Berger-Parker dominance index.
 
     Notes
@@ -51,22 +102,20 @@ def berger_parker_d(counts):
        foraminifera in deep-sea sediments. Science, 168(3937), 1345-1347.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    return counts.max() / N
+    return counts.max() / counts.sum()
 
 
+@_validate_alpha(empty=np.nan)
 def brillouin_d(counts):
     r"""Calculate Brillouin's diversity index.
 
-    Brillouin's diversity index (:math:`H_B`) is defined as
+    Brillouin's diversity index (:math:`H_B`) is defined as:
 
     .. math::
 
-       H_B = \frac{\ln N!-\sum^s_{i=1}{\ln n_i!}}{N}
+       H_B = \frac{\ln N!-\sum_{i=1}^S{\ln n_i!}}{N}
 
-    where :math:`N` is the total number of individuals in the sample, :math:`s`
+    where :math:`N` is the total number of individuals in the sample, :math:`S`
     is the number of taxa, and :math:`n_i` is the number of individuals in the
     :math:`i^{\text{th}}` taxon.
 
@@ -77,7 +126,7 @@ def brillouin_d(counts):
 
     Returns
     -------
-    double
+    float
         Brillouin's diversity index.
 
     Notes
@@ -90,34 +139,45 @@ def brillouin_d(counts):
        Press. New York.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    nz = counts[counts.nonzero()]
-    return (gammaln(N + 1) - gammaln(nz + 1).sum()) / N
+    return (gammaln((N := counts.sum()) + 1) - gammaln(counts + 1).sum()) / N
 
 
-def dominance(counts):
+@_validate_alpha(empty=np.nan)
+def dominance(counts, finite=False):
     r"""Calculate Simpson's dominance index.
 
     Simpson's dominance index, a.k.a. Simpson's :math:`D`, measures the degree
-    of concentration of taxon composition of a sample. It is defined as
+    of concentration of taxon composition of a sample. It is defined as:
 
     .. math::
 
-       D = \sum{p_i^2}
+       D = \sum_{i=1}^S{p_i^2}
 
-    where :math:`p_i` is the proportion of the entire sample that taxon
-    :math:`i` represents.
+    where :math:`S` is the number of taxa and :math:`p_i` is the proportion
+    of the sample represented by taxon :math:`i`.
+
+    Simpson's :math:`D` ranges from 0 (infinite diversity; no dominance) and 1
+    (complete dominance, no diversity).
 
     Simpson's :math:`D` can be interpreted as the probability that two randomly
-    selected individuals belong to the same taxon. It ranges between 0 and 1.
+    selected individuals belong to the same taxon.
+
+    Simpson's :math:`D` may be corrected for finite samples to account for the
+    effect of sampling without replacement. This more accurately represents the
+    above probability when the sample is small. It is calculated as:
+
+    .. math::
+
+       D = \frac{\sum_{i=1}^s{n_i(n_i - 1))}}{N(N - 1)}
+
+    where :math:`n_i` is the number of individuals in the :math:`i^{\text{th}}`
+    taxon and :math:`N` is the total number of individuals in the sample.
 
     Simpson's :math:`D` is sometimes referred to as "Simpson's index". It
     should be noted that :math:`D` is not a measure of community diversity. It
     is also important to distinguish :math:`D` from Simpson's diversity index
-    (:math:`1 - D`) and Simpson's reciprocal index (:math:`1 / D`), both of
-    which are measures of community diversity.
+    (:math:`1 - D`) and inverse Simpson index (:math:`1 / D`), both of which
+    are measures of community diversity.
 
     Discrepancy exists among literature in using the term "Simpson index" and
     the denotion :math:`D`. It is therefore important to distinguish these
@@ -127,10 +187,12 @@ def dominance(counts):
     ----------
     counts : 1-D array_like, int
         Vector of counts.
+    finite : bool, optional
+        If ``True``, correct for finite sampling.
 
     Returns
     -------
-    double
+    float
         Simpson's dominance index.
 
     See Also
@@ -147,12 +209,14 @@ def dominance(counts):
        688-688.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    return ((counts / N) ** 2).sum()
+    if finite:
+        D = (counts * (counts - 1)).sum() / ((N := counts.sum()) * (N - 1))
+    else:
+        D = ((counts / counts.sum()) ** 2).sum()
+    return D
 
 
+@_validate_alpha()
 def doubles(counts):
     """Calculate number of double-occurrence taxa (doubletons).
 
@@ -167,55 +231,64 @@ def doubles(counts):
         Doubleton count.
 
     """
-    counts = _validate_counts_vector(counts)
     return (counts == 2).sum()
 
 
-def enspie(counts):
+def enspie(counts, finite=False):
     r"""Calculate ENS_pie alpha diversity measure.
 
-    ENS_pie is equivalent to ``1 / dominance``:
+    The effective number of species (ENS) derived from Hurlbert's probability
+    of interspecific encounter (PIE) ([1]_, [2]_) is defined as:
 
     .. math::
 
-       ENS_{pie} = \frac{1}{\sum_{i=1}^s{p_i^2}}
+       ENS_{pie} = \frac{1}{\sum_{i=1}^S{p_i^2}}
 
-    where :math:`s` is the number of taxa and :math:`p_i` is the proportion
-    of the sample represented by taxon :math:`i`.
+    where :math:`S` is the number of taxa and :math:`p_i` is the proportion of
+    the sample represented by taxon :math:`i`.
+
+    Therefore, :math:`ENS_{pie}` is equivalent to the inverse Simpson index
+    (``1 / D``).
 
     Parameters
     ----------
     counts : 1-D array_like, int
         Vector of counts.
+    finite : bool, optional
+        If ``True``, correct for finite sampling.
 
     Returns
     -------
-    double
+    float
         ENS_pie alpha diversity measure.
 
     See Also
     --------
+    inv_simpson
     dominance
 
     Notes
     -----
-    ENS_pie is defined in [1]_.
+    ``enspie`` is an alias for ``inv_simpson``.
 
     References
     ----------
-    .. [1] Chase and Knight (2013). "Scale-dependent effect sizes of ecological
-       drivers on biodiversity: why standardised sampling is not enough".
-       Ecology Letters, Volume 16, Issue Supplement s1, pgs 17-26.
+    .. [1] Chase, J. M., & Knight, T. M. (2013). Scale-dependent effect sizes
+       of ecological drivers on biodiversity: why standardised sampling is not
+       enough. Ecology letters, 16, 17-26.
+
+    .. [2] Hurlbert, S. H. (1971). The nonconcept of species diversity: a
+       critique and alternative parameters. Ecology, 52(4), 577-586.
 
     """
-    counts = _validate_counts_vector(counts)
-    return 1 / dominance(counts)
+    return inv_simpson(counts, finite=finite)
 
 
+@_validate_alpha(empty=np.nan)
 def esty_ci(counts):
     r"""Calculate Esty's confidence interval of Good's coverage estimator.
 
-    Esty's confidence interval is defined as
+    Esty's confidence interval is defined as:
 
     .. math::
 
@@ -225,7 +298,7 @@ def esty_ci(counts):
     total number of individuals, and :math:`z` is a constant that depends on
     the targeted confidence and based on the normal distribution.
 
-    :math:`W` is defined as
+    :math:`W` is defined as:
 
     .. math::
 
@@ -243,6 +316,10 @@ def esty_ci(counts):
     tuple
         Esty's confidence interval as ``(lower_bound, upper_bound)``.
 
+    See Also
+    --------
+    goods_coverage
+
     Notes
     -----
     Esty's confidence interval was originally described in [1]_.
@@ -255,19 +332,15 @@ def esty_ci(counts):
        estimator of the coverage of a random sample". Ann Statist 11: 905-912.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-
-    f1 = singles(counts)
-    f2 = doubles(counts)
+    N = counts.sum()
+    f1 = (counts == 1).sum()
+    f2 = (counts == 2).sum()
     z = 1.959963985
     W = (f1 * (N - f1) + 2 * N * f2) / (N**3)
-
     return f1 / N - z * np.sqrt(W), f1 / N + z * np.sqrt(W)
 
 
-@np.errstate(invalid="ignore")
+@_validate_alpha(empty=np.nan)
 def fisher_alpha(counts):
     r"""Calculate Fisher's alpha, a metric of diversity.
 
@@ -288,7 +361,7 @@ def fisher_alpha(counts):
 
     Returns
     -------
-    double
+    float
         Fisher's alpha.
 
     Raises
@@ -317,14 +390,8 @@ def fisher_alpha(counts):
        sample of an animal population. The Journal of Animal Ecology, pp.42-58.
 
     """
-    counts = _validate_counts_vector(counts)
-
-    # alpha = 0 when sample has no individual
-    if (N := counts.sum()) == 0:
-        return 0.0
-
     # alpha = +inf when all taxa are singletons
-    if N == (S := sobs(counts)):
+    if (N := counts.sum()) == (S := counts.size):
         return np.inf
 
     # objective function to minimize:
@@ -333,7 +400,8 @@ def fisher_alpha(counts):
         return (x * np.log(1 + (N / x)) - S) ** 2 if x > 0 else np.inf
 
     # minimize the function using the default method (Brent's algorithm)
-    res = minimize_scalar(f)
+    with np.errstate(invalid="ignore"):
+        res = minimize_scalar(f)
 
     # there is a chance optimization could fail
     if res.success is False:
@@ -342,11 +410,13 @@ def fisher_alpha(counts):
     return res.x
 
 
+@_validate_alpha(empty=np.nan)
 def goods_coverage(counts):
     r"""Calculate Good's coverage estimator.
 
-    Good's coverage estimator :math:`C` is an estimation of the proportion of
-    the population represented in the sample. It is defined as
+    Good's coverage estimator :math:`C`, a.k.a. Turing estimator or Good-
+    Turing (GT) estimator, is an estimation of the proportion of the
+    population represented in the sample. It is defined as:
 
     .. math::
 
@@ -362,8 +432,13 @@ def goods_coverage(counts):
 
     Returns
     -------
-    double
+    float
         Good's coverage estimator.
+
+    See Also
+    --------
+    esty_ci
+    robbins
 
     Notes
     -----
@@ -375,13 +450,10 @@ def goods_coverage(counts):
        estimation of population parameters. Biometrika, 40(3-4), 237-264.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    f1 = singles(counts)
-    return 1 - (f1 / N)
+    return 1 - ((counts == 1).sum() / counts.sum())
 
 
+@_validate_alpha()
 def heip_e(counts):
     r"""Calculate Heip's evenness measure.
 
@@ -391,8 +463,8 @@ def heip_e(counts):
 
        \frac{(e^H-1)}{(S-1)}
 
-    where :math:`H` is the Shannon-Wiener entropy of counts (using logarithm
-    base :math:`e`) and :math:`S` is the number of taxa in the sample.
+    where :math:`H` is Shannon's diversity index and :math:`S` is the number
+    of taxa in the sample.
 
     Parameters
     ----------
@@ -401,7 +473,7 @@ def heip_e(counts):
 
     Returns
     -------
-    double
+    float
         Heip's evenness measure.
 
     See Also
@@ -413,21 +485,44 @@ def heip_e(counts):
     -----
     Heip's evenness measure was originally described in [1]_.
 
+    When there is only one taxon, the return value is 1.0.
+
     References
     ----------
     .. [1] Heip, C. 1974. A new index measuring evenness. J. Mar. Biol. Ass.
        UK., 54, 555-557.
 
     """
-    counts = _validate_counts_vector(counts)
-    return (np.exp(shannon(counts, base=np.e)) - 1) / (sobs(counts) - 1)
+    if (S := counts.size) == 0:
+        return np.nan
+    elif S == 1:
+        return 1.0
+    return (shannon(counts, exp=True) - 1) / (S - 1)
 
 
+@_validate_alpha(empty=np.nan)
 def kempton_taylor_q(counts, lower_quantile=0.25, upper_quantile=0.75):
-    """Calculate Kempton-Taylor Q index of alpha diversity.
+    r"""Calculate Kempton-Taylor Q index of alpha diversity.
 
-    Estimates the slope of the cumulative abundance curve in the interquantile
-    range. By default, uses lower and upper quartiles, rounding inwards.
+    Kempton-Taylor Q index measures diversity based on the middle-ranking taxa
+    in the abundance distribution. Specifically, it estimates the slope of the
+    cumulative abundance curve in the interquantile range. It is defined as:
+
+    .. math::
+
+       Q = \frac{S_{lower..upper}}{\ln n_{lower} - \ln n_{upper}}
+
+    where "lower" and "upper" are the taxa at the lower and upper quantiles of
+    the abundance distribution, :math:`S` is the number of taxa, and :math:`n`
+    is the number of individuals.
+
+    By default, the lower and upper quartiles are used. Therefore:
+
+    .. math::
+
+       Q = \frac{S}{2(\ln n_{0.25} - \ln n_{0.75})}
+
+    The quantiles are rounded inwards in this implementation.
 
     Parameters
     ----------
@@ -440,7 +535,7 @@ def kempton_taylor_q(counts, lower_quantile=0.25, upper_quantile=0.75):
 
     Returns
     -------
-    double
+    float
         Kempton-Taylor Q index of alpha diversity.
 
     Notes
@@ -464,14 +559,59 @@ def kempton_taylor_q(counts, lower_quantile=0.25, upper_quantile=0.75):
     .. [2] http://www.pisces-conservation.com/sdrhelp/index.html
 
     """
-    counts = _validate_counts_vector(counts)
-    n = len(counts)
-    lower = int(np.ceil(n * lower_quantile))
-    upper = int(n * upper_quantile)
+    S = counts.size
+    lower = int(np.ceil(S * lower_quantile))
+    upper = int(S * upper_quantile)
     sorted_counts = np.sort(counts)
     return (upper - lower) / np.log(sorted_counts[upper] / sorted_counts[lower])
 
 
+def inv_simpson(counts, finite=False):
+    r"""Calculate inverse Simpson index.
+
+    The inverse Simpson index (:math:`1 / D`), a.k.a., Simpson's reciprocal
+    index, is defined as:
+
+    .. math::
+
+       1 / D = \frac{1}{\sum_{i=1}^S{p_i^2}}
+
+    where :math:`S` is the number of taxa and :math:`p_i` is the proportion
+    of the sample represented by taxon :math:`i`.
+
+    Parameters
+    ----------
+    counts : 1-D array_like, int
+        Vector of counts.
+    finite : bool, optional
+        If ``True``, correct for finite sampling when calculating :math:`D`.
+
+    Returns
+    -------
+    float
+        Inverse Simpson index.
+
+    See Also
+    --------
+    dominance
+
+    Notes
+    -----
+    :math:`1 / D` is a measurement of the effective number of species (ENS).
+    It is equivalent to Hill number with order 2 (:math:`^2D`).
+
+    Inverse Simpson index was originally described in [1]_.
+
+    References
+    ----------
+    .. [1] Simpson, E. H. (1949). Measurement of diversity. Nature, 163(4148),
+       688-688.
+
+    """
+    return 1 / dominance(counts, finite=finite)
+
+
+@_validate_alpha(empty=np.nan)
 def margalef(counts):
     r"""Calculate Margalef's richness index.
 
@@ -484,7 +624,7 @@ def margalef(counts):
     where :math:`S` is the number of taxa and :math:`N` is the total number
     of individuals in the sample.
 
-    Assumes log accumulation.
+    Margalef's richness index assumes log accumulation.
 
     Parameters
     ----------
@@ -493,8 +633,12 @@ def margalef(counts):
 
     Returns
     -------
-    double
+    float
         Margalef's richness index.
+
+    See Also
+    --------
+    menhinick
 
     Notes
     -----
@@ -506,12 +650,12 @@ def margalef(counts):
        3, 36-71.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    return (sobs(counts) - 1) / np.log(N)
+    if (N := counts.sum()) == 1:
+        return np.nan
+    return (counts.size - 1) / np.log(N)
 
 
+@_validate_alpha(empty=np.nan)
 def mcintosh_d(counts):
     r"""Calculate McIntosh dominance index.
 
@@ -538,7 +682,7 @@ def mcintosh_d(counts):
 
     Returns
     -------
-    double
+    float
         McIntosh dominance index.
 
     See Also
@@ -555,17 +699,17 @@ def mcintosh_d(counts):
        certain concepts to diversity. Ecology 48, 1115-1126.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
+    if (N := counts.sum()) == 1:
+        return np.nan
     u = np.sqrt((counts**2).sum())
     return (N - u) / (N - np.sqrt(N))
 
 
+@_validate_alpha(empty=np.nan)
 def mcintosh_e(counts):
-    r"""Calculate McIntosh's evenness measure E.
+    r"""Calculate McIntosh's evenness measure.
 
-    McIntosh evenness measure E is defined as:
+    McIntosh's evenness measure :math:`E` is defined as:
 
     .. math::
 
@@ -582,8 +726,8 @@ def mcintosh_e(counts):
 
     Returns
     -------
-    double
-        McIntosh evenness measure E.
+    float
+        McIntosh evenness measure.
 
     See Also
     --------
@@ -591,8 +735,7 @@ def mcintosh_e(counts):
 
     Notes
     -----
-    The implementation here is based on the description given in [1]_, **NOT**
-    the one in the SDR-IV online manual, which is wrong.
+    McIntosh's evenness measure was originally described in [1]_.
 
     References
     ----------
@@ -600,14 +743,14 @@ def mcintosh_e(counts):
        Indices. p 560.
 
     """
-    counts = _validate_counts_vector(counts)
-    numerator = np.sqrt((counts * counts).sum())
+    S = counts.size
     N = counts.sum()
-    S = sobs(counts)
+    numerator = np.sqrt((counts * counts).sum())
     denominator = np.sqrt((N - S + 1) ** 2 + S - 1)
     return numerator / denominator
 
 
+@_validate_alpha(empty=np.nan)
 def menhinick(counts):
     r"""Calculate Menhinick's richness index.
 
@@ -620,7 +763,7 @@ def menhinick(counts):
     where :math:`S` is the number of taxa and :math:`N` is the total number
     of individuals in the sample.
 
-    Assumes square-root accumulation.
+    Menhinick's richness index assumes square-root accumulation.
 
     Parameters
     ----------
@@ -629,8 +772,12 @@ def menhinick(counts):
 
     Returns
     -------
-    double
+    float
         Menhinick's richness index.
+
+    See Also
+    --------
+    margalef
 
     Notes
     -----
@@ -642,18 +789,20 @@ def menhinick(counts):
        76-77.
 
     """
-    counts = _validate_counts_vector(counts)
-    return sobs(counts) / np.sqrt(counts.sum())
+    return counts.size / np.sqrt(counts.sum())
 
 
+@_validate_alpha(empty=np.nan)
 def michaelis_menten_fit(counts, num_repeats=1, params_guess=None):
     r"""Calculate Michaelis-Menten fit to rarefaction curve of observed taxa.
 
-    The Michaelis-Menten equation is defined as:
+    The Michaelis-Menten equation estimates the asymptote of the rarefaction
+    curve. It is an estimator of the true richness of a community given the
+    observation. It is defined as:
 
     .. math::
 
-       S=\frac{nS_{max}}{n+B}
+       S = \frac{nS_{max}}{n+B}
 
     where :math:`n` is the number of individuals and :math:`S` is the number of
     taxa. This function estimates the :math:`S_{max}` parameter.
@@ -677,7 +826,7 @@ def michaelis_menten_fit(counts, num_repeats=1, params_guess=None):
 
     Returns
     -------
-    S_max : double
+    float
         Estimate of the :math:`S_{max}` parameter in the Michaelis-Menten
         equation.
 
@@ -699,8 +848,6 @@ def michaelis_menten_fit(counts, num_repeats=1, params_guess=None):
        Michaelis-Menten equation. Biometrics 43, 793-803.
 
     """
-    counts = _validate_counts_vector(counts)
-
     n_indiv = counts.sum()
     if params_guess is None:
         S_max_guess = sobs(counts)
@@ -726,36 +873,6 @@ def michaelis_menten_fit(counts, num_repeats=1, params_guess=None):
     ]
 
 
-def sobs(counts):
-    """Calculate the observed species richness of a sample.
-
-    Observed species richness, usually denoted as :math:`S_{obs}` or simply
-    :math:`S`, is the number of distinct species (i.e., taxa), or any discrete
-    groups of biological entities found in a sample.
-
-    It should be noted that observed species richness is smaller than or equal
-    to the true species richness of a population from which the sample is
-    collected.
-
-    Parameters
-    ----------
-    counts : 1-D array_like, int
-        Vector of counts.
-
-    Returns
-    -------
-    int
-        Observed species richness.
-
-    See Also
-    --------
-    observed_features
-
-    """
-    counts = _validate_counts_vector(counts)
-    return (counts != 0).sum()
-
-
 def observed_features(counts):
     """Calculate the number of distinct features.
 
@@ -775,7 +892,7 @@ def observed_features(counts):
 
     Notes
     -----
-    `observed_features` is an alias for `sobs`.
+    ``observed_features`` is an alias for ``sobs``.
 
     """
     return sobs(counts)
@@ -805,7 +922,7 @@ def observed_otus(counts):
 
     Notes
     -----
-    `observed_otus` is an alias for `sobs`.
+    ``observed_otus`` is an alias for ``sobs``.
 
     """
     # @deprecated
@@ -814,6 +931,7 @@ def observed_otus(counts):
     return sobs(counts)
 
 
+@_validate_alpha()
 def osd(counts):
     """Calculate observed taxa, singletons, and doubletons.
 
@@ -839,19 +957,19 @@ def osd(counts):
     on these three measures.
 
     """
-    counts = _validate_counts_vector(counts)
-    return sobs(counts), singles(counts), doubles(counts)
+    return counts.size, (counts == 1).sum(), (counts == 2).sum()
 
 
-def pielou_e(counts):
+@_validate_alpha()
+def pielou_e(counts, base=None):
     r"""Calculate Pielou's evenness index.
 
     Pielou's evenness index (:math:`J'`), a.k.a., Shannon's equitability index
-    (:math:`E_H`), is defined as
+    (:math:`E_H`), is defined as:
 
     .. math::
 
-       J' = \frac{(H)}{\ln(S)}
+       J' = \frac{(H)}{\log(S)}
 
     where :math:`H` is the Shannon index of the sample and :math:`S` is the
     number of taxa in the sample.
@@ -864,10 +982,12 @@ def pielou_e(counts):
     ----------
     counts : 1-D array_like, int
         Vector of counts.
+    base : int or float, optional
+        Logarithm base to use in the calculation. Default is ``e``.
 
     Returns
     -------
-    double
+    float
         Pielou's evenness index.
 
     See Also
@@ -879,16 +999,26 @@ def pielou_e(counts):
     -----
     Pielou's evenness index was originally described in [1]_.
 
+    When there is only one taxon, the return value is 1.0.
+
     References
     ----------
     .. [1] Pielou, E. C., 1966. The measurement of diversity in different types
        of biological collections. Journal of Theoretical Biology, 13, 131-44.
 
     """
-    counts = _validate_counts_vector(counts)
-    return 0.0 if (H := shannon(counts, base=np.e)) == 0.0 else H / np.log(sobs(counts))
+    if (S := counts.size) == 0:
+        return np.nan
+    elif S == 1:
+        return 1.0
+    H = shannon(counts, base=base)
+    H_max = np.log(S)
+    if base is not None:
+        H_max /= np.log(base)
+    return H / H_max
 
 
+@_validate_alpha(empty=np.nan)
 def robbins(counts):
     r"""Calculate Robbins' estimator for probability of unobserved outcomes.
 
@@ -896,9 +1026,13 @@ def robbins(counts):
 
     .. math::
 
-       \frac{F_1}{n+1}
+       \frac{F_1}{N}
 
-    where :math:`F_1` is the number of singleton taxa.
+    where :math:`F_1` is the number of singleton taxa and and :math:`N` is the
+    total number of individuals in the sample.
+
+    The result can be interpreted as the probability of discovering a new taxon
+    at the :math:`N`-th individual given the current :math:`N - 1` individuals.
 
     Parameters
     ----------
@@ -907,95 +1041,130 @@ def robbins(counts):
 
     Returns
     -------
-    double
-        Robbins' estimate.
+    float
+        Robbins' estimator.
+
+    See Also
+    --------
+    goods_coverage
 
     Notes
     -----
-    Robbins' estimator is defined in [1]_. The estimate computed here is for
-    :math:`n-1` counts, i.e. the x-axis is off by 1.
+    Robbins' estimator is defined in [1]_.
 
     References
     ----------
-    .. [1] Robbins, H. E (1968). Ann. of Stats. Vol 36, pp. 256-257.
+    .. [1] Robbins, H. E. (1968). Estimating the total probability of the
+       unobserved outcomes of an experiment. Ann. Math. Statist., 39(6),
+       256-257.
 
     """
-    counts = _validate_counts_vector(counts)
-    return singles(counts) / counts.sum()
+    return (counts == 1).sum() / counts.sum()
 
 
-def shannon(counts, base=2):
-    r"""Calculate Shannon's diversity index, default in bits.
+@_validate_alpha(empty=np.nan)
+def shannon(counts, base=None, exp=False):
+    r"""Calculate Shannon's diversity index.
 
     Shannon's diversity index, :math:`H'`, a.k.a., Shannon index, or Shannon-
-    Wiener index, is defined as
+    Wiener index, is equivalent to entropy in information theory. It is defined
+    as:
 
     .. math::
 
-       H' = -\sum_{i=1}^s\left(p_i\log_2 p_i\right)
+       H' = -\sum_{i=1}^S\left(p_i\log_b(p_i)\right)
 
-    where :math:`s` is the number of taxa and :math:`p_i` is the proportion
+    where :math:`S` is the number of taxa and :math:`p_i` is the proportion
     of the sample represented by taxon :math:`i`.
+
+    The logarithm base :math:`b` defaults to ``e``, but may be 2, 10 or other
+    custom values.
+
+    The exponential of Shannon index, :math:`exp(H')`, measures the effective
+    number of species (a.k.a., true diversity). It is equivalent to perplexity
+    in information theory, or Hill number with order 1 (:math:`^1D`). The value
+    is independent from the base:
+
+    .. math::
+
+       exp(H') = b ^ {-\sum_{i=1}^S\left(p_i\log_b(p_i)\right)} = \prod_{i=1}
+       ^{S}p_i^{-p_i}
 
     Parameters
     ----------
     counts : 1-D array_like, int
         Vector of counts.
-    base : scalar, optional
-        Logarithm base to use in the calculation.
+    base : int or float, optional
+        Logarithm base to use in the calculation. Default is ``e``.
+    exp : bool, optional
+        If ``True``, return the exponential of Shannon index.
 
     Returns
     -------
-    double
+    float
         Shannon's diversity index.
 
     Notes
     -----
-    Shannon's diversity index was originally proposed in [1]_ as a measure of
-    entropy.
+    Shannon index (i.e., entropy) was originally proposed in [1]_. The
+    exponential of Shannon index (i.e., perplexity) was discussed in [2]_ in
+    the context of community diversity.
 
-    The default logarithm base used here is 2 instead of :math:`e`.
+    .. note:: The default logarithm base was changed from 2 to :math:`e` in
+       version 0.6.1, for consistency with the majority of literature.
 
     References
     ----------
     .. [1] Shannon, C. E. (1948). A mathematical theory of communication. The
        Bell system technical journal, 27(3), 379-423.
 
+    .. [2] Jost, L. (2006). Entropy and diversity. Oikos, 113(2), 363-375.
+
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    freqs = counts / N
-    nonzero_freqs = freqs[freqs.nonzero()]
-    if nonzero_freqs.size <= 1:
-        return 0.0
-    return -(nonzero_freqs * np.log(nonzero_freqs)).sum() / np.log(base)
+    freqs = counts / counts.sum()
+
+    # perplexity
+    if exp is True:
+        return (freqs**-freqs).prod()
+
+    # entropy
+    else:
+        H = (-freqs * np.log(freqs)).sum()
+        if base is not None:
+            H /= np.log(base)
+        return H
 
 
-def simpson(counts):
+def simpson(counts, finite=False):
     r"""Calculate Simpson's diversity index.
 
     Simpson's diversity index, a.k.a., Gini-Simpson index, or Gini impurity,
-    is defined as
+    is defined as:
 
     .. math::
 
-       1 - \sum{p_i^2}
+       1 - \sum_{i=1}^S{p_i^2}
 
-    where :math:`p_i` is the proportion of the sample represented by taxon
-    :math:`i`.
+    where :math:`S` is the number of taxa and :math:`p_i` is the proportion
+    of the sample represented by taxon :math:`i`.
 
     Therefore, Simpson's diversity index is also denoted as :math:`1 - D`, in
     which :math:`D` is the Simpson's dominance index.
+
+    Simpson's diversity index can be interpreted as the probability that two
+    randomly selected individuals belong to different taxa. It is also known
+    as Hurlbert's probability of interspecific encounter (PIE).
 
     Parameters
     ----------
     counts : 1-D array_like, int
         Vector of counts.
+    finite : bool, optional
+        If ``True``, correct for finite sampling when calculating :math:`D`.
 
     Returns
     -------
-    double
+    float
         Simpson's diversity index.
 
     See Also
@@ -1006,20 +1175,54 @@ def simpson(counts):
     -----
     Simpson's diversity index was originally described in [1]_.
 
+    Hurlbert's probability of interspecific encounter was described in [2]_.
+
     References
     ----------
     .. [1] Simpson, E. H. (1949). Measurement of diversity. Nature, 163(4148),
        688-688.
 
+    .. [2] Hurlbert, S. H. (1971). The nonconcept of species diversity: a
+       critique and alternative parameters. Ecology, 52(4), 577-586.
+
     """
-    counts = _validate_counts_vector(counts)
-    return 1 - dominance(counts)
+    return 1 - dominance(counts, finite=finite)
 
 
+def simpson_d(counts, finite=False):
+    """Calculate Simpson's dominance index, a.k.a. Simpson's D.
+
+    Parameters
+    ----------
+    counts : 1-D array_like, int
+        Vector of counts.
+    finite : bool, optional
+        If ``True``, correct for finite sampling.
+
+    Returns
+    -------
+    int
+        Simpson's dominance index.
+
+    See Also
+    --------
+    dominance
+    simpson
+    simpson_e
+
+    Notes
+    -----
+    ``simpson_d`` is an alias for ``dominance``.
+
+    """
+    return dominance(counts, finite=finite)
+
+
+@_validate_alpha(empty=np.nan)
 def simpson_e(counts):
     r"""Calculate Simpson's evenness index.
 
-    Simpson's evenness (a.k.a., equitability) index :math:`E_D` is defined as
+    Simpson's evenness (a.k.a., equitability) index :math:`E_D` is defined as:
 
     .. math::
 
@@ -1040,13 +1243,12 @@ def simpson_e(counts):
 
     Returns
     -------
-    double
+    float
         Simpson's evenness index.
 
     See Also
     --------
     dominance
-    enspie
     simpson
 
     Notes
@@ -1055,15 +1257,22 @@ def simpson_e(counts):
 
     References
     ----------
-    .. [1] Simpson, E. H. (1949). Measurement of diversity. nature, 163(4148), 688-688.
-    .. [2] Pielou, E. C. (1966). The measurement of diversity in different types of
-       biological collections. Journal of theoretical biology, 13, 131-144.
+    .. [1] Simpson, E. H. (1949). Measurement of diversity. nature, 163(4148),
+       688-688.
+
+    .. [2] Pielou, E. C. (1966). The measurement of diversity in different
+       types of biological collections. Journal of theoretical biology, 13,
+       131-144.
 
     """
-    counts = _validate_counts_vector(counts)
-    return enspie(counts) / sobs(counts)
+    # Note: the finite version of simpson_e might be: 1 / (D(S + 1)), because
+    # S + 1 is the maximum possible finite D given S. Otherwise, the result can
+    # be greater than 1 for small samples. However, I didn't find literature
+    # stating this. Therefore, the `finite` parameter is not used here.
+    return 1 / (counts.size * dominance(counts))
 
 
+@_validate_alpha()
 def singles(counts):
     """Calculate number of single-occurrence taxa (singletons).
 
@@ -1078,10 +1287,40 @@ def singles(counts):
         Singleton count.
 
     """
-    counts = _validate_counts_vector(counts)
     return (counts == 1).sum()
 
 
+@_validate_alpha()
+def sobs(counts):
+    """Calculate the observed species richness of a sample.
+
+    Observed species richness, usually denoted as :math:`S_{obs}` or simply
+    :math:`S`, is the number of distinct species (i.e., taxa), or any discrete
+    groups of biological entities found in a sample.
+
+    It should be noted that observed species richness is smaller than or equal
+    to the true species richness of a population from which the sample is
+    collected.
+
+    Parameters
+    ----------
+    counts : 1-D array_like, int
+        Vector of counts.
+
+    Returns
+    -------
+    int
+        Observed species richness.
+
+    See Also
+    --------
+    observed_features
+
+    """
+    return counts.size
+
+
+@_validate_alpha(empty=np.nan)
 def strong(counts):
     r"""Calculate Strong's dominance index.
 
@@ -1105,7 +1344,7 @@ def strong(counts):
 
     Returns
     -------
-    double
+    float
         Strong's dominance index.
 
     Notes
@@ -1118,10 +1357,7 @@ def strong(counts):
        and between plant communities. Community Ecology, 3, 237-246.
 
     """
-    counts = _validate_counts_vector(counts)
-    if (N := counts.sum()) == 0:
-        return 0.0
-    S = sobs(counts)
-    i = np.arange(1, len(counts) + 1)
+    S = counts.size
     sorted_sum = np.sort(counts)[::-1].cumsum()
-    return (sorted_sum / N - (i / S)).max()
+    i = np.arange(1, S + 1)
+    return (sorted_sum / counts.sum() - (i / S)).max()