+++ skbio/diversity/alpha/tests/test_base.py
@@ -8,17 +8,17 @@
 
 from unittest import TestCase, main
 from io import StringIO
+import warnings
 
 import numpy as np
 import numpy.testing as npt
 
 from skbio import TreeNode
 from skbio.diversity.alpha import (
-    berger_parker_d, brillouin_d, dominance, doubles, enspie,
-    esty_ci, fisher_alpha, goods_coverage, heip_e, kempton_taylor_q,
-    margalef, mcintosh_d, mcintosh_e, menhinick, michaelis_menten_fit,
-    observed_features, observed_otus, osd, pielou_e, robbins, shannon,
-    simpson, simpson_e, singles, sobs, strong)
+    berger_parker_d, brillouin_d, dominance, doubles, enspie, esty_ci, fisher_alpha,
+    goods_coverage, heip_e, inv_simpson, kempton_taylor_q, margalef, mcintosh_d,
+    mcintosh_e, menhinick, michaelis_menten_fit, observed_features, observed_otus, osd,
+    pielou_e, robbins, shannon, simpson, simpson_d, simpson_e, singles, sobs, strong)
 
 
 class BaseTests(TestCase):
@@ -39,44 +39,40 @@ class BaseTests(TestCase):
         self.assertEqual(berger_parker_d(np.array([1, 1, 1, 1, 0])), 0.25)
         self.assertEqual(berger_parker_d(self.counts), 5 / 22)
         self.assertEqual(berger_parker_d(np.array([5])), 1)
-        self.assertEqual(berger_parker_d([0, 0, 0]), 0)
-        self.assertEqual(berger_parker_d([]), 0)
+        self.assertTrue(np.isnan(berger_parker_d([0, 0])))
 
     def test_brillouin_d(self):
         self.assertAlmostEqual(brillouin_d(np.array([1, 2, 0, 0, 3, 1])),
                                0.86289353018248782)
-        self.assertEqual(brillouin_d([0, 0, 0]), 0)
-        self.assertEqual(brillouin_d([]), 0)
+        self.assertTrue(np.isnan(brillouin_d([0, 0])))
 
     def test_dominance(self):
         self.assertEqual(dominance(np.array([5])), 1)
         self.assertAlmostEqual(dominance(np.array([1, 0, 2, 5, 2])), 0.34)
-        self.assertEqual(dominance([0, 0, 0]), 0)
-        self.assertEqual(dominance([]), 0)
+        self.assertTrue(np.isnan(dominance([0, 0])))
+
+        # finite sample correction
+        self.assertEqual(dominance(np.array([5]), finite=True), 1)
+        self.assertAlmostEqual(dominance(
+            np.array([1, 0, 2, 5, 2]), finite=True), 0.8 / 3)
+        self.assertTrue(np.isnan(dominance([0, 0], finite=True)))
 
     def test_doubles(self):
         self.assertEqual(doubles(self.counts), 3)
         self.assertEqual(doubles(np.array([0, 3, 4])), 0)
         self.assertEqual(doubles(np.array([2])), 1)
-        self.assertEqual(doubles(np.array([0, 0])), 0)
+        self.assertEqual(doubles([0, 0]), 0)
 
     def test_enspie(self):
-        # Totally even community should have ENS_pie = number of taxa.
-        self.assertAlmostEqual(enspie(np.array([1, 1, 1, 1, 1, 1])), 6)
-        self.assertAlmostEqual(enspie(np.array([13, 13, 13, 13])), 4)
-
-        # Hand calculated.
-        arr = np.array([1, 41, 0, 0, 12, 13])
-        exp = 1 / ((arr / arr.sum()) ** 2).sum()
-        self.assertAlmostEqual(enspie(arr), exp)
-
-        # Using dominance.
-        exp = 1 / dominance(arr)
-        self.assertAlmostEqual(enspie(arr), exp)
-
-        arr = np.array([1, 0, 2, 5, 2])
-        exp = 1 / dominance(arr)
-        self.assertAlmostEqual(enspie(arr), exp)
+        for vec in (
+            np.array([1, 1, 1, 1, 1, 1]),
+            np.array([1, 41, 0, 0, 12, 13]),
+            np.array([1, 0, 2, 5, 2])
+        ):
+            self.assertEqual(enspie(vec), inv_simpson(vec))
+        vec = np.array([1, 2, 3, 4])
+        self.assertEqual(enspie(vec, finite=True),
+                         inv_simpson(vec, finite=True))
 
     def test_esty_ci(self):
         def _diversity(indices, f):
@@ -112,6 +108,8 @@ class BaseTests(TestCase):
         npt.assert_array_almost_equal(observed_lower, expected_lower)
         npt.assert_array_almost_equal(observed_upper, expected_upper)
 
+        self.assertTrue(np.isnan(esty_ci([0, 0])))
+
     def test_fisher_alpha(self):
         exp = 2.7823796
         arr = np.array([4, 3, 4, 0, 1, 0, 2])
@@ -137,10 +135,6 @@ class BaseTests(TestCase):
         obs = fisher_alpha([1, 1, 1, 1, 1])
         self.assertEqual(obs, np.inf)
 
-        # Should be zero when there is no individual
-        obs = fisher_alpha([0, 0, 0, 0, 0])
-        self.assertEqual(obs, 0)
-
         # Should be large when most species are singletons
         obs = fisher_alpha([1] * 99 + [2])
         self.assertAlmostEqual(obs, 5033.278, places=3)
@@ -149,24 +143,54 @@ class BaseTests(TestCase):
         obs = fisher_alpha([1] * 999 + [2])
         TestCase().assertAlmostEqual(obs, 500333.3, places=1)
 
+        self.assertTrue(np.isnan(fisher_alpha([0, 0])))
+
     def test_goods_coverage(self):
         counts = [1] * 75 + [2, 2, 2, 2, 2, 2, 3, 4, 4]
         obs = goods_coverage(counts)
         self.assertAlmostEqual(obs, 0.23469387755)
-        self.assertEqual(goods_coverage([0, 0, 0]), 0)
-        self.assertEqual(goods_coverage([]), 0)
+        self.assertTrue(np.isnan(goods_coverage([0, 0])))
 
     def test_heip_e(self):
         # Calculate "by hand".
         arr = np.array([1, 2, 3, 1])
-        h = shannon(arr, base=np.e)
-        expected = (np.exp(h) - 1) / 3
+        H = shannon(arr)
+        expected = (np.exp(H) - 1) / (arr.size - 1)
         self.assertEqual(heip_e(arr), expected)
 
         # From Statistical Ecology: A Primer in Methods and Computing, page 94,
         # table 8.1.
         self.assertAlmostEqual(heip_e([500, 300, 200]), 0.90, places=2)
         self.assertAlmostEqual(heip_e([500, 299, 200, 1]), 0.61, places=2)
+
+        # Edge cases
+        self.assertEqual(heip_e([5]), 1)
+        self.assertTrue(np.isnan(heip_e([0, 0])))
+
+    def test_inv_simpson(self):
+        # Totally even community should have 1 / D = number of taxa.
+        self.assertAlmostEqual(inv_simpson(np.array([1, 1, 1, 1, 1, 1])), 6)
+        self.assertAlmostEqual(inv_simpson(np.array([13, 13, 13, 13])), 4)
+
+        # Hand calculated.
+        arr = np.array([1, 41, 0, 0, 12, 13])
+        exp = 1 / ((arr / arr.sum()) ** 2).sum()
+        self.assertAlmostEqual(inv_simpson(arr), exp)
+
+        # Using dominance.
+        exp = 1 / dominance(arr)
+        self.assertAlmostEqual(inv_simpson(arr), exp)
+
+        arr = np.array([1, 0, 2, 5, 2])
+        exp = 1 / dominance(arr)
+        self.assertAlmostEqual(inv_simpson(arr), exp)
+
+        # Finite sample correction.
+        self.assertEqual(inv_simpson(
+            np.array([1, 0, 2, 5, 2]), finite=True), 3.75)
+        self.assertEqual(inv_simpson(np.array([3, 3, 3]), finite=True), 4)
+
+        self.assertTrue(np.isnan(inv_simpson([0, 0])))
 
     def test_kempton_taylor_q(self):
         # Approximate Magurran 1998 calculation p143.
@@ -179,26 +203,30 @@ class BaseTests(TestCase):
         np.random.shuffle(arr)
         self.assertAlmostEqual(kempton_taylor_q(arr), exp)
 
+        self.assertTrue(np.isnan(kempton_taylor_q([0, 0])))
+
     def test_margalef(self):
         self.assertEqual(margalef(self.counts), 8 / np.log(22))
-        self.assertEqual(margalef([0, 0, 0]), 0)
-        self.assertEqual(margalef([]), 0)
+        self.assertTrue(np.isnan(margalef([1])))
+        self.assertTrue(np.isnan(margalef([0, 0])))
 
     def test_mcintosh_d(self):
         self.assertAlmostEqual(mcintosh_d(np.array([1, 2, 3])),
                                0.636061424871458)
-        self.assertEqual(mcintosh_d([0, 0, 0]), 0)
-        self.assertEqual(mcintosh_d([]), 0)
+        self.assertTrue(np.isnan(mcintosh_d([1])))
+        self.assertTrue(np.isnan(mcintosh_d([0, 0])))
 
     def test_mcintosh_e(self):
         num = np.sqrt(15)
         den = np.sqrt(19)
         exp = num / den
         self.assertEqual(mcintosh_e(np.array([1, 2, 3, 1])), exp)
+        self.assertTrue(np.isnan(mcintosh_e([0, 0])))
 
     def test_menhinick(self):
         # observed species richness = 9, total # of individuals = 22
         self.assertEqual(menhinick(self.counts), 9 / np.sqrt(22))
+        self.assertTrue(np.isnan(menhinick([0, 0])))
 
     def test_michaelis_menten_fit(self):
         obs = michaelis_menten_fit([22])
@@ -220,17 +248,17 @@ class BaseTests(TestCase):
         # [0,2,4,6] looks like 3 taxa with maybe more to be found.
         self.assertTrue(obs_few > obs_many)
 
+        self.assertTrue(np.isnan(michaelis_menten_fit([0, 0])))
+
     def test_observed_features(self):
-        for obs in [np.array([4, 3, 4, 0, 1, 0, 2]),
-                    np.array([0, 0, 0]),
-                    self.counts]:
-            self.assertEqual(observed_features(obs), sobs(obs))
+        for vec in (np.array([4, 3, 4, 0, 1, 0, 2]), self.counts):
+            self.assertEqual(observed_features(vec), sobs(vec))
 
     def test_observed_otus(self):
-        for obs in [np.array([4, 3, 4, 0, 1, 0, 2]),
-                    np.array([0, 0, 0]),
-                    self.counts]:
-            self.assertEqual(observed_otus(obs), sobs(obs))
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            for vec in (np.array([4, 3, 4, 0, 1, 0, 2]), self.counts):
+                self.assertEqual(observed_otus(vec), sobs(vec))
 
     def test_osd(self):
         self.assertEqual(osd(self.counts), (9, 3, 3))
@@ -238,10 +266,14 @@ class BaseTests(TestCase):
     def test_pielou_e(self):
         # Calculate "by hand".
         arr = np.array([1, 2, 3, 1])
-        h = shannon(arr, np.e)
-        s = 4
-        expected = h / np.log(s)
+        H = shannon(arr)
+        S = arr.size
+        expected = H / np.log(S)
         self.assertAlmostEqual(pielou_e(arr), expected)
+
+        # alternative logarithm base
+        expected = shannon(arr, base=2) / np.log2(S)
+        self.assertAlmostEqual(pielou_e(arr, base=2), expected)
 
         self.assertAlmostEqual(pielou_e(self.counts), 0.92485490560)
 
@@ -252,23 +284,51 @@ class BaseTests(TestCase):
         # Examples from
         # http://ww2.mdsg.umd.edu/interactive_lessons/biofilm/diverse.htm#3
         self.assertAlmostEqual(pielou_e([1, 1, 196, 1, 1]), 0.078, 3)
-        self.assertEqual(pielou_e([0, 0, 200, 0, 0]), 0)
-        self.assertEqual(pielou_e([0, 0, 0, 0, 0]), 0)
-        self.assertEqual(pielou_e([]), 0)
+
+        # Edge cases
+        self.assertEqual(pielou_e([5]), 1)
+        self.assertTrue(np.isnan(pielou_e([0, 0])))
 
     def test_robbins(self):
         self.assertEqual(robbins(np.array([1, 2, 3, 0, 1])), 2 / 7)
+        self.assertTrue(np.isnan(robbins([0, 0])))
 
     def test_shannon(self):
-        self.assertEqual(shannon(np.array([])), 0)
+        self.assertAlmostEqual(shannon([5, 5]), 0.693147181)
+        self.assertEqual(shannon([5, 5], base=2), 1)
+        self.assertAlmostEqual(shannon([5, 5], base=10), 0.301029996)
+
+        # taxa with 0 counts are excluded from calculation
+        self.assertAlmostEqual(shannon([1, 2, 3, 4]), 1.279854226)
+        self.assertAlmostEqual(shannon([0, 1, 2, 3, 4]), 1.279854226)
+
+        # Shannon index of a single-taxon community is always 0
         self.assertEqual(shannon(np.array([5])), 0)
-        self.assertEqual(shannon(np.array([5, 5])), 1)
-        self.assertEqual(shannon(np.array([0, 0, 0])), 0)
-        self.assertEqual(shannon(np.array([1, 1, 1, 1, 0])), 2)
+
+        # Shannon index cannot be calculated for an empty community
+        self.assertTrue(np.isnan(shannon([0, 0])))
+
+        # NaN still holds if input is empty (instead of 0's), this behavior is
+        # different from scipy.stats.entropy, which would return 0.0.
+        self.assertTrue(np.isnan(shannon([])))
+
+        # Exponential of Shannon index
+        self.assertAlmostEqual(shannon([1, 2, 3, 4], exp=True), 3.596115467)
+
+        # Equally abundant taxa, exp(H) = # taxa
+        self.assertEqual(shannon([5, 5, 5], exp=True), 3.0)
 
     def test_simpson(self):
         self.assertAlmostEqual(simpson(np.array([1, 0, 2, 5, 2])), 0.66)
-        self.assertAlmostEqual(simpson(np.array([5])), 0)
+        self.assertEqual(simpson(np.array([5])), 0)
+        self.assertEqual(simpson(np.array([5]), finite=True), 0)
+        self.assertTrue(np.isnan(simpson([0, 0])))
+
+    def  test_simpson_d(self):
+        for vec in (np.array([5]), np.array([1, 0, 2, 5, 2])):
+            self.assertEqual(dominance(vec), simpson_d(vec))
+            self.assertEqual(dominance(vec, finite=True),
+                             simpson_d(vec, finite=True))
 
     def test_simpson_e(self):
         # A totally even community should have simpson_e = 1.
@@ -289,11 +349,13 @@ class BaseTests(TestCase):
         exp = 1 / (D * 4)
         self.assertEqual(simpson_e(arr), exp)
 
+        self.assertTrue(np.isnan(simpson_e([0, 0])))
+
     def test_singles(self):
         self.assertEqual(singles(self.counts), 3)
         self.assertEqual(singles(np.array([0, 3, 4])), 0)
         self.assertEqual(singles(np.array([1])), 1)
-        self.assertEqual(singles(np.array([0, 0])), 0)
+        self.assertEqual(singles([0, 0]), 0)
 
     def test_sobs(self):
         obs = sobs(np.array([4, 3, 4, 0, 1, 0, 2]))
@@ -307,8 +369,7 @@ class BaseTests(TestCase):
 
     def test_strong(self):
         self.assertAlmostEqual(strong(np.array([1, 2, 3, 1])), 0.214285714)
-        self.assertEqual(strong([0, 0, 0]), 0)
-        self.assertEqual(strong([]), 0)
+        self.assertTrue(np.isnan(strong([0, 0])))
 
 
 if __name__ == '__main__':